🧪 added options test
